Switch serialization test to use json path tests instead of verify
