--- test/Sentry.Tests/SerializationTests.Serialization_TransactionAndSpanData.verified.txt
@@ -1,44 +1,0 @@
-﻿{
-  EventId: {},
-  SpanId: {},
-  ParentSpanId: {},
-  Name: name,
-  Platform: csharp,
-  StartTimestamp: DateTimeOffset_1,
-  Operation: operation,
-  Description: ,
-  IsSampled: false,
-  SampleRate: 0.0,
-  Request: {},
-  Contexts: {
-    trace: {
-      SpanId: {},
-      ParentSpanId: {},
-      Operation: operation,
-      Description: ,
-      IsSampled: false,
-      Extra: {
-        transaction1: transaction_value
-      }
-    }
-  },
-  User: {},
-  Sdk: {},
-  Extra: {
-    transaction1: transaction_value
-  },
-  Spans: [
-    {
-      SpanId: {},
-      ParentSpanId: {},
-      StartTimestamp: DateTimeOffset_2,
-      IsFinished: false,
-      Operation: childop,
-      IsSampled: false,
-      Extra: {
-        span1: value1
-      }
-    }
-  ],
-  IsFinished: false
-}
